+++ examples/plot_minimal_pydeseq2_pipeline.py
@@ -15,8 +15,8 @@ We start by importing required packages and setting up an optional path to save 
 import os
 import pickle as pkl
 
-from pydeseq2.DeseqDataSet import DeseqDataSet
-from pydeseq2.DeseqStats import DeseqStats
+from pydeseq2.dds import DeseqDataSet
+from pydeseq2.ds import DeseqStats
 from pydeseq2.utils import load_example_data
 
 SAVE = False  # whether to save the outputs of this notebook
